+++ src/components/Cards/Cards.module.css
@@ -1,0 +1,93 @@
+#event-arrow {
+  height: 10%;
+  padding-right: 25px;
+  width: 10%;
+}
+
+.event-card {
+  position: relative;
+}
+
+.event-content {
+  background: #353535;
+  border-radius: 20px;
+  color: white;
+  cursor: pointer;
+  font-size: 1.2em;
+  font-weight: 300;
+  line-height: 1.4;
+  margin: 0 0 35px 22.5%;
+  padding: 2em 2em 2.5em 2em;
+  position: relative;
+  width: 60%;
+}
+
+.event-content h3 {
+  color: white;
+  line-height: 50px;
+  padding: 5px 60px 0 30px;
+}
+
+.event-content span {
+  font-size: 1em !important;
+  padding: 0 0 0 30px;
+}
+
+.event-content span span {
+  padding: 0 0 0 10px;
+}
+
+.event-icon {
+  background: #353535;
+  border-radius: 50%;
+  box-shadow: 0 0 0 8px #2de1c2;
+  color: #fff;
+  line-height: 40px;
+  font-size: 1.4em;
+  font-style: normal;
+  font-variant: normal;
+  font-weight: normal;
+  height: 40px;
+  left: 18%;
+  margin: 0 0 -3px -25px;
+  position: absolute;
+  speak: none;
+  text-transform: none;
+  width: 40px;
+  -webkit-font-smoothing: antialiased;
+  text-align: center;
+  top: 40%;
+}
+
+.event-icon img {
+  height: 30px;
+  margin-bottom: 3px;
+  width: 30px;
+}
+
+/* .event-link {
+  background: #2de1c2;
+  border-top-left-radius: 90px;
+  border-top-right-radius: 90px;
+  bottom: 0;
+  height: 45px;
+  left: 45%;
+  position: absolute;
+  width: 90px;
+} */
+
+.event-link span {
+  height: 30px;
+  width: 30px;
+}
+
+.event-tag {
+  background: #1ab89d;
+  border-radius: 10px;
+  display: inline-block;
+  padding: 0 15px 0 15px;
+}
+
+.event-tags {
+  padding: 20px 0 0 25px;
+}
