+++ src/components/Cards/index.js
@@ -1,0 +1,3 @@
+import EventCard from "./EventCard.jsx";
+
+export { EventCard };
