--- src/components/Modals/index.js
@@ -1,3 +1,0 @@
-import EventModal from "./EventModal";
-
-export { EventModal };
