+++ src/components/Timeline/Timeline.jsx
@@ -1,23 +1,21 @@
 import React from "react";
 import FireStoreParser from "firestore-parser";
 
+import EventCard from "../Cards/EventCard";
 import { firestoreURL } from "../../constants/constants";
 import styles from "./Timeline.module.css";
-import { EventModal } from "../Modals";
-import EventCard from "./EventCard";
 
 class Timeline extends React.PureComponent {
   constructor(props) {
     super(props);
     this.state = {
-      events: [],
-      selectedEvent: null
+      events: []
     };
   }
 
+  // Gather the events, preformat the data, and sort by startDate
   componentWillMount() {
     let events = [];
-
     fetch(`${firestoreURL}events`)
       .then(response => response.json())
       .then(json => FireStoreParser(json))
@@ -29,46 +27,27 @@ class Timeline extends React.PureComponent {
           event.endDate = new Date(event.endDate);
           events.push(event);
         });
-
         events.sort(function compare(a, b) {
           return a.startDate - b.startDate;
         });
-
         this.setState({ events });
       });
   }
 
-  renderEvents = () => {
-    let oThis = this;
-    let events = [];
+  // renderEvents builds an EventCard for each event defined in the state
+  renderEvents = events => {
     let i = 0;
-
-    this.state.events.forEach(function(event) {
-      events.push(
-        <EventCard key={i} event={event} onClick={oThis.toggleEvent} />
-      );
+    let renderedEvents = [];
+    events.forEach(function(event) {
+      renderedEvents.push(<EventCard key={i} {...event} />);
       i++;
     });
-    return events;
-  };
-
-  toggleEvent = event => {
-    this.setState({ selectedEvent: event });
+    return renderedEvents;
   };
 
   render() {
-    const { selectedEvent } = this.state;
-
-    return (
-      <div className={styles["timeline-container"]}>
-        <ul className={styles["timeline"]}>{this.renderEvents()}</ul>
-        {selectedEvent ? (
-          <EventModal event={selectedEvent} onClose={this.toggleEvent} />
-        ) : (
-          <div />
-        )}
-      </div>
-    );
+    const { events } = this.state;
+    return <ul className={styles["timeline"]}>{this.renderEvents(events)}</ul>;
   }
 }
