+++ src/components/Timeline/Timeline.module.css
@@ -1,7 +1,3 @@
-.timeline-container {
-  height: 100%;
-}
-
 .timeline {
   list-style: none;
   padding: 50px 0 0 0;
@@ -12,135 +8,9 @@
   background: #2de1c2;
   bottom: 0;
   content: "";
-  left: 20%;
+  left: 18%;
   margin-left: -10px;
   position: absolute;
-  top: 50px;
+  top: 30%;
   width: 10px;
-}
-
-.timeline > li {
-  position: relative;
-}
-
-.event-time {
-  display: block;
-  padding-right: 8%;
-  position: absolute;
-  width: 25%;
-}
-
-.event-time span {
-  color: white;
-  display: block;
-  text-align: right;
-}
-
-.event-time span:first-child {
-  font-size: 2.5em;
-}
-
-.event-time span:last-child {
-  font-size: 1.5em;
-}
-
-.event-icon {
-  background: #353535;
-  border-radius: 50%;
-  box-shadow: 0 0 0 8px #2de1c2;
-  color: #fff;
-  line-height: 40px;
-  font-size: 1.4em;
-  font-style: normal;
-  font-variant: normal;
-  font-weight: normal;
-  height: 40px;
-  left: 20%;
-  margin: 0 0 -3px -25px;
-  position: absolute;
-  speak: none;
-  text-transform: none;
-  width: 40px;
-  -webkit-font-smoothing: antialiased;
-  text-align: center;
-  top: 8px;
-}
-
-.event-icon img {
-  height: 30px;
-  margin-bottom: 3px;
-  width: 30px;
-}
-
-.event-content {
-  background: #353535;
-  border-radius: 5px 10px 10px 10px;
-  box-shadow: 0 8px 5px grey;
-  color: white;
-  cursor: pointer;
-  font-size: 1.2em;
-  font-weight: 300;
-  line-height: 1.4;
-  margin: 0 0 14px 25%;
-  padding: 2em;
-  position: relative;
-  width: 60%;
-}
-
-.event-content:after {
-  border: solid transparent;
-  border-right-color: #353535;
-  border-width: 10px;
-  content: " ";
-  height: 0;
-  pointer-events: none;
-  position: absolute;
-  right: 100%;
-  width: 0;
-  top: 4px;
-}
-
-.event-content h4 {
-  color: white;
-  line-height: 50px;
-  padding: 5px 60px 0 30px;
-}
-
-.event-content span {
-  font-size: 0.9em !important;
-  padding: 0 0 0 30px;
-}
-
-.event-content span span {
-  padding: 0 0 0 10px;
-}
-
-.event-link {
-  align-items: center;
-  background: #2de1c2;
-  border-radius: 0 10px 10px 0;
-  display: flex;
-  height: 100%;
-  justify-content: center;
-  position: absolute;
-  right: 0;
-  top: 0;
-  width: 7%;
-}
-
-#event-arrow {
-  padding-right: 25px;
-}
-
-.event-tags {
-  padding: 0 0 0 30px;
-}
-
-.event-tag {
-  background: #7f8691;
-  border-radius: 20px;
-  display: inline-block;
-  font-size: 1em;
-  padding: 5px 15px 5px 15px;
-  margin: 20px 20px 0 0;
 }
